- Reordered member declaration of double blood meal parameters for clarity, as well as use in gui. - Added Mosquitoes.h and Mosquitoes.cpp to encapsulate adult female cohorts in densim code. - Converted densim simulation to use double precision instead of single/float precision. - Began work on NewMosquitoLifeCycle which will use new cohort classes, tracking age and weight. - Parceled out NewMosquitoLifeCycle code into separate methods, AdvanceSusceptibleNulliparous and Advance SusceptibleParous, to reduce method size. - Added member for minimum oviposition temperature from Biology. - Removed used of PreOviAdultCohort and OviAdultCohortCollection from cimsim, using new AdultCohort class from Mosquitoes.h. - Created various collections for new mosquito cohorts. - Added temporary verification methods that output cohorts. - Fixed bug where dsport::DBloodUProp, which is expected to be the proportion at the low weight, or the "upper proportion" (misnomer), was being loaded with the HighWeightRatio instead of LowWeightRatio, from Biology. - Removed extraneous check for DBloodUWt - DBloodLWt == 0 in MosquitoLifeCycle(). - Began work on infection of mosquitoes using new cohorts. - Added helper methods for new cohorts, to sum, calculate average weight, find biting counts, etc.

+++ trunk/simulation/dsport/Humans.h
@@ -34,12 +34,12 @@ public:
 public:
   int FirstDay;
   int LastDay;
-  float Proportion;             // proportion of population in age class
-  float YearlyDeathRate;        // yearly death rate per 1000 for age class
-  float YearlyBirthRate;        // yearly birth rate per 1000 for age class
+  double Proportion;             // proportion of population in age class
+  double YearlyDeathRate;        // yearly death rate per 1000 for age class
+  double YearlyBirthRate;        // yearly birth rate per 1000 for age class
 
-  float DailyDeathRate;
-  float DailyBirthRate;
+  double DailyDeathRate;
+  double DailyBirthRate;
 };
 
 
@@ -78,8 +78,8 @@ class HumanPopulation
 {
 // Typedefs
 public:
-  typedef std::map<int,std::map<int,float>> ClassSpecificSeroprevalence;
-  typedef std::map<int,std::map<int,float>> DailySequentialInfections;
+  typedef std::map<int,std::map<int,double>> ClassSpecificSeroprevalence;
+  typedef std::map<int,std::map<int,double>> DailySequentialInfections;
 
 // Constructors
 public:
@@ -94,13 +94,13 @@ public:
   int GetPopulationSize(void);
 
   std::vector<int> GetInitialAgeDistribution(void);
-  std::vector<float> GetInitialAgeProportions(void);
+  std::vector<double> GetInitialAgeProportions(void);
   std::vector<int> GetAgeDistribution(void);
-  std::vector<float> GetAgeProportions(void);
+  std::vector<double> GetAgeProportions(void);
   std::vector<int> GetBirthsByClass(void);
-  std::vector<float> GetBirthPercentagesByClass(void);
+  std::vector<double> GetBirthPercentagesByClass(void);
   std::vector<int> GetDeathsByClass(void);
-  std::vector<float> GetDeathPercentagesByClass(void);
+  std::vector<double> GetDeathPercentagesByClass(void);
 
   std::vector<std::vector<int>> GetInitialSeroDistribution(void);
   std::vector<std::vector<int>> GetSeroDistribution(void);
@@ -155,12 +155,12 @@ private:
   std::vector<int> _initialAgeDistribution;
   std::vector<int> _ageDistribution;
 
-  std::vector<float> _cumulativeDeaths;
-  std::vector<float> _cumulativeBirths;
+  std::vector<double> _cumulativeDeaths;
+  std::vector<double> _cumulativeBirths;
   std::vector<int> _totalDeaths;
   std::vector<int> _totalBirths;
 
-  std::vector<std::vector<float>> _initialSeroprevalence;
+  std::vector<std::vector<double>> _initialSeroprevalence;
   std::vector<std::vector<int>> _seroDistribution;
   std::vector<std::vector<int>> _initialSeroDistribution;
 
@@ -188,7 +188,7 @@ private:
 
   DailySequentialInfections _dailySequentialInfections;
 
-  std::vector<float> _cumulativeHfDeaths;
+  std::vector<double> _cumulativeHfDeaths;
 };
 
 };
